Rework spline/shape/geometry representation and definition. Add MeshExporter class.
Handling of action space and action handling is now in the geometry class while shape/spline definitions provide the action space informations. This makes it possible to later on add geometries
which can handle depended action easier and also action which move multiple points/have perodicities and through them can reduce the DOF. Addidtionally it cleans up the parser_environment file a lot.

Addidtionally a MeshExporter class is added to make it easier to handle different exports and conditional exports.